requirements.txt
requests
PyGObject
PyYAML